// payload-types.ts
export interface Page {
  id: string;
  title: string;
  hero: {
    type: 'highImpact' | 'mediumImpact' | 'lowImpact';
    richText: {
      [k: string]: unknown;
    }[];
    links: {
      link: {
        type?: 'reference' | 'custom';
        newTab?: boolean;
        reference: {
          value: string | Page;
          relationTo: 'pages';
        };
        url: string;
        label: string;
        appearance?: 'default' | 'primary' | 'secondary';
      };
      id?: string;
    }[];
    media: string | Media;
  };
  layout: (
    | {
      bgColor?: 'white' | 'black';
      richText: {
        [k: string]: unknown;
      }[];
      links: {
        link: {
          type?: 'reference' | 'custom';
          newTab?: boolean;
          reference: {
            value: string | Page;
            relationTo: 'pages';
          };
          url: string;
          label: string;
        };
        id?: string;
      }[];
      id?: string;
      blockName?: string;
      blockType: 'cta';
    }
    | {
      bgColor?: 'white' | 'black';
      layout?: 'oneColumn' | 'twoThirdsOneThird' | 'halfAndHalf' | 'threeColumns';
      columnOne: {
        richText: {
          [k: string]: unknown;
        }[];
        enableLink?: boolean;
        link: {
          type?: 'reference' | 'custom';
          newTab?: boolean;
          reference: {
            value: string | Page;
            relationTo: 'pages';
          };
          url: string;
          label: string;
          appearance?: 'default' | 'primary' | 'secondary';
        };
      };
      columnTwo: {
        richText: {
          [k: string]: unknown;
        }[];
        enableLink?: boolean;
        link: {
          type?: 'reference' | 'custom';
          newTab?: boolean;
          reference: {
            value: string | Page;
            relationTo: 'pages';
          };
          url: string;
          label: string;
          appearance?: 'default' | 'primary' | 'secondary';
        };
      };
      columnThree: {
        richText: {
          [k: string]: unknown;
        }[];
        enableLink?: boolean;
        link: {
          type?: 'reference' | 'custom';
          newTab?: boolean;
          reference: {
            value: string | Page;
            relationTo: 'pages';
          };
          url: string;
          label: string;
          appearance?: 'default' | 'primary' | 'secondary';
        };
      };
      id?: string;
      blockName?: string;
      blockType: 'content';
    }
    | {
      bgColor?: 'white' | 'black';
      position?: 'default' | 'fullscreen';
      media: string | Media;
      caption?: {
        [k: string]: unknown;
      }[];
      id?: string;
      blockName?: string;
      blockType: 'mediaBlock';
    }
  )[];
  slug?: string;
  meta: {
    title?: string;
    description?: string;
    image?: string | Media;
  };
  _status?: 'draft' | 'published';
  createdAt: string;
  updatedAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "media".
 */
export interface Media {
  id: string;
  alt: string;
  url?: string;
  filename?: string;
  mimeType?: string;
  filesize?: number;
  width?: number;
  height?: number;
  createdAt: string;
  updatedAt: string;
}